% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superApply.R
\name{clean_interruption}
\alias{clean_interruption}
\title{Helper of superApply}
\usage{
clean_interruption(jobArray, workingDir, idPrefix)
}
\description{
Executes cleaning code after user sends cleaning signal (ctrl+c, ESC)
}
