% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superApply.R
\name{createStringFunction}
\alias{createStringFunction}
\title{Helper of superApply}
\usage{
createStringFunction(fun, inside = NULL)
}
\arguments{
\item{fun}{character - the fucntion name as a string}

\item{inside}{character vector - the items to put inside function}
}
\value{
character vector - of the form [1]"fun(inside[1])" ... [n]"fun(inside[n]))" where n is the lengh of inside. If inside is NULL it returns an empty string
}
\description{
Takes a function name and a character vector to be put inside
independent calls of the function.
}
\examples{
createStringFunction("library", c("ggplot2", "dyplr")) 
#[1] "library(ggplot2)" "library(dyplr)"

createStringFunction("library") 
#[1] ""
}
