% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superApply.R
\name{getJobArray}
\alias{getJobArray}
\title{Helper of superApply}
\usage{
getJobArray(x, FUN, ..., idPrefix, partitionIndeces, workingDir,
  extraScriptLines, extraBashLines, JobArrayPars, packages, sources)
}
\arguments{
\item{x}{list/vector - data to be partition}

\item{FUN}{function - function to be applied to each element of x}

\item{...}{- further arguments of FUN}

\item{idPrefix}{character - prefix for job names}

\item{partitionIndeces}{list - output of getPartitionIndeces()}
}
\value{
a JobArray object
}
\description{
Submits multiple jobs from the partions of x created in get Partition Indeces
}
