% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getJobNumber.R
\name{getJobNumber}
\alias{getJobNumber}
\title{getJobNumber}
\usage{
getJobNumber(user = Sys.info()["user"])
}
\arguments{
\item{user}{Character - username, if blank it is gotten from sytem info}
}
\value{
Integer - the current number of SLURM jobs in queue
}
\description{
Returns the number of existing SLURM jobs of a given user
}
