% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superApply.R
\name{getUserFunctions}
\alias{getUserFunctions}
\title{Helper of superApply}
\usage{
getUserFunctions()
}
\value{
a character vector with the names of the functions in the global enviroment
}
\description{
Helper of superApply
}
