% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superApply.R
\name{getUserPackages}
\alias{getUserPackages}
\title{Helper of superApply}
\usage{
getUserPackages()
}
\value{
Returns a character vector with the names of the packages in the global enviroment
}
\description{
Helper of superApply
}
