% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grepTempFile}
\alias{grepTempFile}
\title{grepTempFile}
\usage{
grepTempFile(x, pattern, tempLocation = ".")
}
\arguments{
\item{x}{String - path to file}

\item{pattern}{Character vector - patterns to select in file}
}
\description{
Creates a new file based of x only with lines containing
the specified pattern(s)
WARNING: uses unix grep
}
\examples{
grepTempFile("path/to/file", c("patternA", "patternB"))
}
