% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobArray.R
\docType{class}
\name{JobArray}
\alias{JobArray}
\title{Array of Job}
\format{\code{\link{R6Class}} object.}
\usage{
JobArray
}
\value{
Object of \code{\link{R6Class}} with methods for communication with lightning-viz server.
}
\description{
Array of Job
}
\section{Fields}{

\describe{
\item{\code{serveraddress}}{Stores address of your lightning server.}

\item{\code{sessionid}}{Stores id of your current session on the server.}

\item{\code{url}}{Stores url of the last visualization created by this object.}

\item{\code{autoopen}}{Checks if the server is automatically opening the visualizations.}
}}

\examples{
Lightning$new("http://localhost:3000/")
Lightning$new("http://your-lightning.herokuapp.com/")
}
\keyword{data}
