% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobInfo.R
\docType{class}
\name{JobInfo}
\alias{JobInfo}
\title{Class providing a basic plataform to store the options of a SLURM jobs
Other more complex classes like Job and JobArray are children of this class}
\format{\code{\link{R6Class}} object.}
\usage{
JobInfo
}
\value{
Object of \code{\link{R6Class}} with methods for communication with lightning-viz server.
}
\description{
Class providing a basic plataform to store the options of a SLURM jobs
Other more complex classes like Job and JobArray are children of this class
}
\section{Fields}{

\describe{
\item{\code{serveraddress}}{Stores address of your lightning server.}

\item{\code{sessionid}}{Stores id of your current session on the server.}

\item{\code{url}}{Stores url of the last visualization created by this object.}

\item{\code{autoopen}}{Checks if the server is automatically opening the visualizations.}
}}

\examples{
Lightning$new("http://localhost:3000/")
Lightning$new("http://your-lightning.herokuapp.com/")
}
\keyword{data}
