% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superApply.R
\name{mergeListDir}
\alias{mergeListDir}
\title{Helper of superApply}
\usage{
mergeListDir(files, varNames, workingDir)
}
\arguments{
\item{files}{character vector - files to be merged}

\item{varNames}{character vector - varnames associated to each file to be merged}

\item{workingDir}{character - working directory}
}
\description{
merges the result of independ jobs after completion of parallel lapply executions
}
