% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{rSubmitterOpts}
\alias{rSubmitterOpts}
\title{option enviroment}
\format{An object of class \code{environment} of length 4.}
\usage{
rSubmitterOpts
}
\description{
option enviroment
}
\keyword{datasets}
