% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shellCmdExists}
\alias{shellCmdExists}
\title{shellCmdExists 
Internal function for .onAttach Check if shell command exists}
\usage{
shellCmdExists(cmd)
}
\description{
shellCmdExists 
Internal function for .onAttach Check if shell command exists
}
