% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trim}
\alias{trim}
\title{Trims white space}
\usage{
trim(x)
}
\arguments{
\item{x}{String or character vector}
}
\value{
Returns x without any white spaces
}
\description{
Trims white space from a character vector
}
\examples{
trim("A B")
}
