% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getMaxJobArrayLength}
\alias{getMaxJobArrayLength}
\title{getMaxJobArrayLength
Internal function for .onAttach that get the maximum length for a Job array
from the SLURM config file}
\usage{
getMaxJobArrayLength()
}
\description{
getMaxJobArrayLength
Internal function for .onAttach that get the maximum length for a Job array
from the SLURM config file
}
