% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMaxJobArrayLength}
\alias{getMaxJobArrayLength}
\title{Gets maximum length for a Job array in the current SLURM cluster}
\usage{
getMaxJobArrayLength()
}
\description{
Internal function for .onAttach that gets the maximum length for a Job array
from the SLURM config file
}
