% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superApply.R
\name{getUserFunctions}
\alias{getUserFunctions}
\title{getUserFunctions
Helper of superApply}
\usage{
getUserFunctions()
}
\description{
Returns a character vector with the names of the functions
in the global enviroment
}
