% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superApply.R
\name{getUserPackages}
\alias{getUserPackages}
\title{getUserPackages
Helper of superApply}
\usage{
getUserPackages()
}
\description{
Returns a character vector with the names of the functions
in the global enviroment
}
