% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Job.R
\docType{class}
\name{Job}
\alias{Job}
\title{Class providing object with methods for communication with lightning-viz server}
\format{\code{\link{R6Class}} object.}
\usage{
Job
}
\arguments{
\item{removeScript}{Logical - if True the bash script use to submit thorugh sbatch is deleted}
}
\value{
Object of \code{\link{R6Class}} with methods for communication with lightning-viz server.

job - returns self invisibly for method concatenation 
Gets state of job
Waits for job to finish
}
\description{
Class providing object with methods for communication with lightning-viz server
}
\section{Fields}{

\describe{
\item{\code{serveraddress}}{Stores address of your lightning server.}

\item{\code{sessionid}}{Stores id of your current session on the server.}

\item{\code{url}}{Stores url of the last visualization created by this object.}

\item{\code{autoopen}}{Checks if the server is automatically opening the visualizations.}
}}

\examples{
Lightning$new("http://localhost:3000/")
Lightning$new("http://your-lightning.herokuapp.com/")
}
\keyword{data}
