% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobInfo.R
\docType{class}
\name{JobInfo}
\alias{JobInfo}
\title{Parent class of Job and JobArray}
\format{R6 class}
\usage{
JobInfo
}
\value{
\code{\link{R6Class}} with methods and fields for SLURM job manipulation
}
\description{
R6 Class providing an essential platform to process SLURM jobs.
This class should not be explicitly used by the user.
To create, manipulate, and submit SLURM jobs please refer to the Job and JobArray classes
}
\section{Usage}{

\code{
x <- JobInfo$new(jobName = NULL, outDir = NULL, partition = NULL, time = NULL, mem = NULL, proc = NULL, totalProc = NULL, nodes = NULL, email = NULL) \cr
x$wait(stopIfFailed = F, verbose = T) \cr
x$cancel() \cr
x$getState(simplify = F) \cr
x$clean() \cr
}
}

\section{Method description}{

\enumerate{
 \item{**Initialize**}{
     \cr
     \code{x <- JobInfo$new(jobName = NULL, outDir = NULL, partition = NULL, time = NULL, mem = NULL, proc = NULL, totalProc = NULL, nodes = NULL, email = NULL)}
     \cr Parameters:
      \itemize{
         \item{jobName} {: character - Name of job, if NULL one will be generated of the form rSubmitter_job_[random_alphanumeric]. Equivalent to \code{--job-name} of SLURM sbatch. Most output files use it as a suffix}
         \item{outDir} {: character - writeable path for the sabtch script as well as the SLRUM STDERR and STDOUT files. If NULL the current working directory will be used}
         \item{partition} {: character - Partition to use. Equivalent to \code{--partition} of SLURM sbatch}
         \item{time} {: character - Time requested for job execution, one accepted format is "HH:MM:SS". Equivalent to \code{--time} of SLURM sbatch}
         \item{mem} {: character - Memory requested for job execution, one accepted format is "xG" or "xMB". Equivalent to \code{--mem} of SLURM sbatch}
         \item{proc} {: integer - Number of processors requested per task. Equivalent to \code{--cpus-per-task} of SLURM sbatch}
         \item{totalProc} {: integer - Number of tasks requested for job. Equivalent to \code{--ntasks} of SLURM sbatch}
         \item{nodes} {: integer - Number of nodes requested for job. Equivalent to \code{--nodes} of SLURM sbatch}
         \item{email} {: character - email address to send info when job is done. Equivalent to \code{--nodes} of SLURM sbatch}
         }
     \cr Return: \cr object of class \code{Job}
     }
     
 \item{Wait for job(s) to finish}{
     \cr
     \code{x$wait(stopIfFailed = F, verbose = T)}
     \cr Time between each job state check is defined in the entry TIME_WAIT_JOB_STATUS:seconds in the config file located at ~/.rSubmitter
     \cr Parameters:
      \itemize{
         \item{stopIfFailed} {: logical -  if TRUE stops when one job has failed (only useful for JobArray) it then cancels the 
                                 rest of the pending and running jobs. If FALSE and one or more Jobs failed it raises a warning for each failed job}
         \item{verbose} {: logical -  if TRUE prints the job state(s) at every check}
         }
     \cr Return: \cr self - for method concatenation
     }

 \item{Cancel job(s)}{
     \cr
     \code{x$cancel()}
     \cr Return: \cr self - for method concatenation
     }

 \item{Get job(s) state}{
     \cr
     \code{x$getState(simplify = F)}
     \cr Parameters:
      \itemize{
         \item{simplify} {: logical - if TRUE returns a freqeuncy data.frame of job states, otherwise returns individual jobs and their associated job names, job ids, and states}
         }
     \cr Return: \cr data.frame - With SLURM states
     }

 \item{Remove SLURM-associated files}{
     \cr
     \code{x$clean(script = TRUE, out = TRUE, err = TRUE)}
     \cr Parameters:
      \itemize{
         \item{script} {: logical - if TRUE deletes sbatch submission script(s) associated to this object}
         \item{out} {: logical - if TRUE deletes STDOUT file(s) from SLURM associated to this object} 
         \item{err} {: logical - if TRUE deletes STDERR file(s) from SLURM associated to this object}
         }
     \cr Return: \cr self - for method concatenation
     }
 }
}

\examples{
\dontrun{
jobInfo <- JobInfo$new()
}
}
\keyword{datasets}
