% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printTime}
\alias{printTime}
\title{Prints message with current time and date}
\usage{
printTime(..., carriageReturn = F)
}
\arguments{
\item{carriageReturn}{Logical, if TRUE print a carriage return at the end of line}

\item{x}{String}
}
\description{
Prints to console the current date and time followed by
the message(s) in x
}
\examples{
printTime(": is right now")
}
