% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superApply.R
\name{superApply}
\alias{superApply}
\title{supperApply, it mimics the functionality of lapply but implemented
in a way that each iteration of the apply is submmitted as an individual
job to a SLURM cluster. Hence emulating a parellel behaivor in apply}
\usage{
superApply(x, FUN, ..., tasks = 1, workingDir, packages = NULL,
  sources = NULL, extraBashLines = "", extraScriptLines = "", clean = F,
  partition = NULL, time = NULL, mem = NULL)
}
\description{
Each job batch, err, out, and script files are stored in a temporary folder. Once
all jobs have been submmitted, the function waits for them to finish. Once the jobs
are done it compiles all the results into list and returns them, therefore fully
mimicking the apply behaivor.
Author: Pablo Garcia
}
