% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systemSubmit.R
\name{systemSubmit}
\alias{systemSubmit}
\title{systemSubmit}
\usage{
systemSubmit(command, n = 12, wait = 5, ignore.stdout = TRUE,
  ignore.stderr = F, intern = TRUE, ...)
}
\arguments{
\item{command}{Character - system command to execute}

\item{n}{Integer - number of times to try executing command in case it returns a non-zero exit status}

\item{wait}{Integer - time in seconds to wait before trying executing the command again}

\item{ignore.stdout}{Logical - if TRUE it won't return the standard output of execution}

\item{ignore.stderr}{Logical - if TRUE it won't return the standard error of execution}

\item{...}{to be passed to system()}
}
\value{
system return object - see ?system
}
\description{
Internal function for the rSubmitter package
Tries executing a command up to n times while the execution returns a
non-zero exit status.
Useful when trying to automatically submmit a job and the scheduler has problems,
then this will try several times if submission fails for any reason
}
